feat: major refactor and re-write
The whole thing has been re-written.
The app no longer relies on helm charts but instead has the resources and workloads baked in.
This allows for a smoother upgrade process where by it doesn't have to be updated in step with the helm chart.

It also allows for direct interaction with the client-go library without much need for obfuscation of the code which can cause confusion.
We still require the Resource struct to allow easier wrapping and grouping of all resources meaning we can run similar functions but the whole process has been smoothed.
The process for creating, validating and testing and deletion has been separated out to allow for more control over resources. this allows for troubleshooting to be done before any deleting is carried out should a test fail.

This commit is to store what has been changed so far with many more commits to come as the test suite is re-written and improved.