Feature release -
The whole thing has been re-written.
The app no longer relies on helm charts but instead has the resources and workloads baked in.
This allows for a smoother upgrade process where by it doesn't have to be updated in step with the helm chart.

It still requires the Resource struct to allow easier wrapping and grouping of all resources meaning we can run similar functions but the whole process has been smoothed.

The process for creating, validating and testing and deletion has been separated out into validation and deletion to allow for more control over resources. This allows for troubleshooting to be done before any deleting is carried out should a test
fail.

GPU testing has been added to the tests.

Monitoring resources has not been included for this release nor has Istio but they will come along with telemetry and metrics.